+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Allow provider `base_url`s to be overriden in config
+
 ## [1.0.0-rc.8] - 2025-10-29
 
 ### Added